+++ pkgs/test/CHANGELOG.md
@@ -3,6 +3,8 @@
 * Ignore an error locating the SDK directory on platforms where the
   `resolvedExecutable` is unexpectedly `null`.
 * Fix a bug where `-c exe` tests would hang on exit on windows.
+* Fix a bug where the web test runner would not find custom HTML files when
+  using the dart2wasm compiler.
 
 ## 1.31.0